requirements.txt
moviepy
pydub
pyyaml